ubusd: use size_t for string and blob length variables
strlen() and blob_raw_len() return size_t; storing their result in
an int relies on an implicit narrowing conversion. Use size_t for
the affected length variables in ubusd_alloc_event_pattern(),
ubus_create_obj_method() and __ubusd_handle_lookup().

ubusd_alloc_event_pattern() checked the length with "<= 0", which is
only ever true for 0 once the variable is unsigned; replace it with
"!len".

Link: https://github.com/openwrt/ubus/pull/20

+++ ubusd_obj.c
@@ -37,7 +37,7 @@ static void ubus_unref_object_type(struct ubus_object_type *type)
 static bool ubus_create_obj_method(struct ubus_object_type *type, struct blob_attr *attr)
 {
 	struct ubus_method *m;
-	int bloblen = blob_raw_len(attr);
+	size_t bloblen = blob_raw_len(attr);
 
 	m = calloc(1, sizeof(*m) + bloblen);
 	if (!m)
